+++ templates/emails/customer-payment-failed.php
@@ -37,7 +37,7 @@ do_action('woocommerce_email_header', $email_heading, $email);
                                         </tr>
                                         <tr>
                                             <td align="center" valign="middle" class="font-primary font-FFFFFF font-16 font-weight-600 pb-5 font-space-0">
-                                                <?php echo __('Payment Failed', 'bypierofracasso-woocommerce-emails'); ?>
+                                                <?php echo __('Payment Failed', 'piero-fracasso-emails'); ?>
                                             </td>
                                         </tr>
                                         <tr>
@@ -51,7 +51,7 @@ do_action('woocommerce_email_header', $email_heading, $email);
                                                     <tr>
                                                         <td align="center" class="bg-FFFFFF block btn border-radius-4">
                                                             <a href="<?php echo ($order instanceof WC_Order) ? esc_url($order->get_checkout_payment_url()) : '#'; ?>" class="font-primary font-4B7BEC font-14 font-weight-600 font-space-0-5 block btn white-space">
-                                                                <?php echo __('TRY AGAIN', 'bypierofracasso-woocommerce-emails'); ?>
+                                                                <?php echo __('TRY AGAIN', 'piero-fracasso-emails'); ?>
                                                             </a>
                                                         </td>
                                                     </tr>
@@ -129,9 +129,9 @@ do_action('woocommerce_email_header', $email_heading, $email);
                                             <td align="left" valign="middle" class="center-text font-primary font-191919 font-18 font-weight-600 font-space-0 pb-20">
                                                 <?php
                                                 if ($order instanceof WC_Order) {
-                                                    echo __('Hello ' . esc_html($order->get_billing_first_name()) . ',', 'bypierofracasso-woocommerce-emails');
+                                                    echo __('Hello ' . esc_html($order->get_billing_first_name()) . ',', 'piero-fracasso-emails');
                                                 } else {
-                                                    echo __('Hello Customer,', 'bypierofracasso-woocommerce-emails');
+                                                    echo __('Hello Customer,', 'piero-fracasso-emails');
                                                 }
                                                 ?>
                                             </td>
@@ -139,7 +139,7 @@ do_action('woocommerce_email_header', $email_heading, $email);
                                         <tr>
                                             <td align="left" valign="middle" class="center-text font-primary font-595959 font-16 font-weight-400 font-space-0 pb-20" style="padding:0px;">
                                                 <?php
-                                                echo __('Unfortunately, we couldn\'t complete your order due to an issue with your payment method.', 'bypierofracasso-woocommerce-emails');
+                                                echo __('Unfortunately, we couldn\'t complete your order due to an issue with your payment method.', 'piero-fracasso-emails');
                                                 if ($additional_content) {
                                                     echo '<br><br>' . wp_kses_post(wptexturize($additional_content));
                                                 }
@@ -150,7 +150,7 @@ do_action('woocommerce_email_header', $email_heading, $email);
                                                 ?>
                                                 <br><br>
                                                 <a href="<?php echo ($order instanceof WC_Order) ? esc_url($order->get_checkout_payment_url()) : '#'; ?>" class="font-4B7BEC">
-                                                    <?php echo __('Try a different payment method', 'bypierofracasso-woocommerce-emails'); ?>
+                                                    <?php echo __('Try a different payment method', 'piero-fracasso-emails'); ?>
                                                 </a>
                                             </td>
                                         </tr>
@@ -237,7 +237,7 @@ do_action('woocommerce_email_order_meta', $order, $sent_to_admin, $plain_text, $
                                                     <tr>
                                                         <td align="center" class="bg-4B7BEC block btn border-radius-4">
                                                             <a href="<?php echo ($order instanceof WC_Order) ? esc_url($order->get_checkout_payment_url()) : '#'; ?>" class="font-primary font-FFFFFF font-14 font-weight-600 font-space-0-5 block btn white-space">
-                                                                <?php echo __('TRY AGAIN', 'bypierofracasso-woocommerce-emails'); ?>
+                                                                <?php echo __('TRY AGAIN', 'piero-fracasso-emails'); ?>
                                                             </a>
                                                         </td>
                                                     </tr>